+++ app/src/main/java/net/invictusmanagement/invictuskiosk/domain/model/Unit.kt
@@ -1,15 +1,15 @@
 package net.invictusmanagement.invictuskiosk.domain.model
 
 data class Unit(
-    val area: Int,
+    val area: Float,
     val availableDateUtc: String,
     val bathrooms: Float,
-    val bedrooms: Int,
+    val bedrooms: Float,
     val floor: Int,
     val id: Int,
     val imageIds: List<Any>,
     val isRentHide: Boolean,
     val isUnitsForSale: Boolean,
-    val rent: Int,
+    val rent: Float,
     val unitNbr: String
 )
